Trying to implement send confirmation email and forgot password function

+++ guestbook.py
@@ -1,4 +1,4 @@
-from os import abort
+from threading import Thread
 
 from flask import Flask, render_template, request, redirect, url_for, jsonify, session, flash
 from flask_sqlalchemy import SQLAlchemy
@@ -8,10 +8,11 @@ import flask_login as fl
 from flask_wtf import FlaskForm
 from flask_wtf.csrf import CSRFProtect
 from wtforms import StringField, PasswordField, SubmitField
-# insantiate Flask
 from sqlalchemy import text
 from wtforms.validators import DataRequired, EqualTo, ValidationError, Email
 from functools import wraps
+from itsdangerous import URLSafeSerializer
+from flask_mail import Mail, Message
 
 sess = Session()
 app = Flask(__name__)
@@ -22,9 +23,18 @@ app.config['SQL_ALCHEMY_TRACK_MODIFICATIONS'] = False
 app.config['SESSION_TYPE'] = 'memcached'
 app.config['SECRET_KEY'] = 'super secret key'
 app.config['SESSION_PERMANENT'] = True
+app.config['MAIL_SERVER'] = 'smtp.gmail.com'
+app.config['MAIL_PORT'] = 465
+app.config['MAIL_USE_TLS'] = False
+app.config['MAIL_USE_SSL'] = True
+app.config['MAIL_USERNAME'] = 'jimath3@gmail.com'  # enter your email here
+app.config['MAIL_DEFAULT_SENDER'] = 'jimath3@gmail.com' # enter your email here
+app.config['MAIL_PASSWORD'] = 'SeGaMaWpOuStH@@GaMwToSpItIsOu@@' # enter your password here
 login_manager = fl.LoginManager()
 login_manager.init_app(app)
 csrf = CSRFProtect(app)
+ts = URLSafeSerializer(app.config["SECRET_KEY"])
+mail = Mail(app)
 
 # instantiate the database (?)
 db = SQLAlchemy(app)
@@ -50,10 +60,11 @@ class User(db.Model, fl.UserMixin):
 	username = db.Column(db.String(20))
 	password = db.Column(db.String(100))
 	email = db.Column(db.String(100))
+	email_confirmed = db.Column(db.Boolean(), default=False)
 
 	def __init__(self, username, password, email):
 		self.username = username
-		self.password = password
+		self.password = self.set_password(password)
 		self.email = email
 		self.password_hash = None
 		# self.is_authenticated = False
@@ -71,6 +82,9 @@ class User(db.Model, fl.UserMixin):
 		self.username = username
 		return self.username'''
 
+class AnonymousUser(fl.AnonymousUserMixin):
+	def __init__(self):
+		self.current_user = "anonymous"
 
 class LoginForm(FlaskForm):
 	username = StringField('Username', validators=[DataRequired()])
@@ -107,23 +121,27 @@ class RegistrationForm(FlaskForm):
 # custom decorator gia na dw an kapoios einai syndedemenos
 def login_required(function):
 	@wraps(function)
-	def wrapper():
+	def wrapper(*args, **kwargs):
 		form = LoginForm()
 		if not fl.current_user.is_authenticated:
 			flash("You do not have permission to view this page", "warning")
 			# abort()
-		return render_template("login.html", form=form)
+			#return render_template("login.html", form=form)
+			return redirect(url_for("login"))
+		return function(*args, **kwargs)
 	return wrapper
 
 # custom decorator gia na dw an o xrhsths pou tsilimpourdizei einai o admin
 def admin_required(function):
 	@wraps(function)
-	def wrapper():
+	def wrapper(*args, **kwargs):
 		form = LoginForm()
-		if not fl.current_user.is_authenticated or fl.current_user.username == 'admin':
-			flash("You do not have permission to view this page", "warning")
+		if not fl.current_user.username == 'admin':
+			flash("You must be admin", "warning")
 			# abort()
-		return render_template("login.html", form=form)
+			return redirect(url_for("login"))
+		#return render_template("login.html")
+		return function(*args, **kwargs)
 	return wrapper
 
 
@@ -176,6 +194,7 @@ def user_loader(id):
 	return User.query.get(str(id))
 
 
+# apla den mou douleve swsta kai to ekana sxolio
 '''@login_manager.request_loader
 def request_loader(request):
 	username = request.form.get('username')
@@ -216,8 +235,6 @@ def login_process():
 	flash("Successful login", "success")
 
 	return redirect(url_for("index"))
-	'''else:
-		flash("Invalid credentials", "error")'''
 
 
 # return redirect(url_for("login"))
@@ -261,9 +278,6 @@ def sign_up_process():
 	user = User(username=username, password=password_hash)
 	db.session.add(user)
 	db.session.commit()
-	session['logged_in'] = True
-	session['username'] = username
-	session['password'] = password_hash
 
 	return redirect(url_for('index'))
 
@@ -279,9 +293,52 @@ def register():
 		user.set_password(form.password.data)
 		db.session.add(user)
 		db.session.commit()
+
+		subject = "Confirm your email"
+		token = ts.dumps(user.email, salt="email-confirm-key")
+		confirm_url = url_for(
+			'send_email',
+			token=token,
+			_external= True
+		)
+
+		html = render_template(
+			"confirm_email.html",
+			confirm_url=confirm_url
+		)
+
+		#apla ypothetw oti exw orisei mia function send_email
+		send_email(user.email, subject, html)
+
 		flash('Congratulations, you are now a registered user!')
-		return redirect(url_for('login'))
+		return redirect(url_for('confirmation'))
 	return render_template('register.html', title='Register', form=form)
+
+#to ftiaxnw gia na exei asynchronh apostolh email mesw thread kai na mhn kathysterhsei to loading ths selidas-stoxoy
+def async_send_email(app, msg):
+	with app.app_context():
+		mail.send(msg)
+
+
+@app.route("/send_email", methods=["GET", "POST"])
+def send_email(user_email, subject, html):
+	#email = ts.loads(token, salt="email-confirm-key", max_age=86400)
+	#user = User.query.filter_by(email=email).first_or_404()
+	#user.email_confirmed = True
+	#db.session.add(user)
+	#db.session.commit()
+	msg = Message(subject, recipients=user_email)
+	msg.body = html
+	'''msg.html = html
+	# mail.send(msg)'''
+	thr = Thread(target = async_send_email, args=[app, msg])
+	flash("A confirmation link was sent to your inbox", "success")
+
+	return thr
+
+@app.route("/confirmation", methods=['GET', 'POST'])
+def confirmation():
+	return render_template("confirm_email.html")
 
 
 @app.route("/logout")
@@ -289,14 +346,15 @@ def logout():
 	fl.logout_user()
 	return redirect(url_for("index"))
 
-	return redirect(url_for('index'))
-
 
 @app.route('/edit_com/', methods=['GET', 'POST'])
 @login_required
 def edit_com():
 	form = EditComForm()
 	com_id = request.args.get('com_id')
+	if com_id is None:
+		flash("There was no comment id", "error")
+		return redirect(url_for('index'))
 	session['com_id'] = com_id
 	query = text("select * from comments where id=" + str(com_id))
 	result = db.engine.execute(query)
@@ -312,6 +370,9 @@ def edit_com():
 def edit_com_process():
 	# com_id = request.form['com_id']
 	com_id = request.form['com_id']
+	if com_id is None:
+		flash("There was no comment id", "error")
+		return redirect(url_for("index"))
 	com_cont = request.form['comment']
 
 	query = text("update comments set comment = '" + str(com_cont) + "' where id = " + str(com_id))
@@ -323,6 +384,8 @@ def edit_com_process():
 
 # made with redirect method
 @app.route('/delete_red/', methods=['GET', 'POST'])
+@login_required
+@admin_required
 def delete_red():
 	# i should use AJAX so that i do not redirect
 	if request.method == 'POST':
@@ -338,8 +401,12 @@ def delete_red():
 
 # uses AJAX to delete
 @app.route('/delete_ajax', methods=['GET', 'POST'])
+@login_required
 def delete_ajax():
 	com_id = request.args.get('com_id')
+	if com_id is None:
+		flash("No comment id provided", "error")
+		return redirect(url_for("index"))
 	query = "delete from comments where id=" + str(com_id)
 	session.pop("comment", None)
 	db.engine.execute(query)
@@ -355,7 +422,10 @@ def delete_ajax():
 @admin_required
 def delete_user_ajax():
 	u_id = request.args.get('u_id')
-	query = "delete from users where id=" + str(u_id)
+	if u_id is None:
+		flash("No user id provided", "error")
+		return redirect(url_for("index"))
+	query = "delete from user where id=" + str(u_id)
 	db.engine.execute(query)
 	db.session.commit()
 
@@ -370,10 +440,6 @@ def users_list():
 		if fl.current_user.username == 'admin':
 			# return redirect(url_for('users_list'))
 			return display_users()
-		'''else:
-			return render_template('you_are_not_admin.html')
-	else:
-		return render_template("login_first.html")'''
 
 
 @app.route("/display_users", methods=['GET', 'POST'])
@@ -393,23 +459,11 @@ def display_users():
 @app.route("/profile/", methods=['GET', 'POST'])
 @login_required
 def profile():
-	if session.get('username') is not None:
-		username = request.args.get('user')
-		query = "select * from users where username='" + str(username) + "'"
-		db.engine.execute(query)
-		db.session.commit()
+	query = "select * from user where username='" + str(fl.current_user.username) + "'"
+	db.engine.execute(query)
+	db.session.commit()
 
-		return render_template("profile.html", result=query)
-	else:
-		return render_template("login_first.html")
-
-
-@app.route('/home', methods=['GET', 'POST'])
-def home():
-	links = ['https://www.google.gr', 'https://www.youtube.com', 'https://www.facebook.com', 'https://www.twitter.com']
-	# it searches for a file in the default path "./templates/FILE_NAME" to open it
-	# the render_template function does the searching job
-	return render_template('example.html', links=links)
+	return render_template("profile.html",  result=query)
 
 
 # <VARIABLE_NAME> is a variable that is passed through the URL bar
